bought-a-plotter/requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.25.0
packaging==23.1
Pillow==9.5.0
pyparsing==3.1.0
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
